refactor: move prompt utils to template module
